refactor: move useEffect content and initialValues to separate file
