+++ src/components/Forms/Annocement/types.ts
@@ -1,4 +1,5 @@
 import { FormikErrors, FormikTouched } from 'formik';
+import * as H from 'history';
 
 export interface RouteParams {
   addId: string;
@@ -22,3 +23,12 @@ export interface FieldsProps {
   errors: FormikErrors<Initial>;
   variant?: 'standard' | 'filled' | 'outlined';
 }
+
+export type RouteTypeFunc = (
+  pathname: string,
+  initialValues: Initial,
+  params: RouteParams,
+  userInfo: { email: string; name: string; phonenumber: string },
+  setInitialValues: React.Dispatch<React.SetStateAction<Initial>>,
+  history: H.History<unknown>
+) => void;
